Get the API Name from a variable, instead of assuming we can guess it
In some cases it is "stage-service" (e.g. "dev-my-api"), but in other
cases it is "service-stage" (e.g. "my-api-dev"). Rather than making
assumptions, just have the code using this module provide us with the
API Name. This also required asking for the desired subdomain
separately, since previously we just used the app name for that, and
this change replaces that `app_name` variable with the `api_name` and
`subdomain` variables.